Extract util function for getting failed invoices
