+++ src/features/recurringInvoices/RecurringInvoicesContainer.tsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import { useMutation, useQuery } from '@apollo/react-hooks';
 import { useTranslation } from 'react-i18next';
 
-import RecurringInvoices, { FailedInvoices } from './RecurringInvoices';
+import RecurringInvoices from './RecurringInvoices';
 import Modal from '../../common/modal/Modal';
 import RecurringInvoicesForm from './recurringInvoicesForm/RecurringInvoicesForm';
 import { RECURRING_INVOICES } from './__generated__/RECURRING_INVOICES';
@@ -14,7 +14,7 @@ import {
 import { SEND_EXISTING_BERTH_INVOICES_MUTATION } from './mutations';
 import authService from '../auth/authService';
 import hdsToast from '../../common/toast/hdsToast';
-import { DataSummaryProps } from '../../common/dataSummary/DataSummary';
+import { getFailedInvoicesData } from './utils';
 
 const RecurringInvoicesContainer = () => {
   const { t } = useTranslation();
@@ -50,42 +50,23 @@ const RecurringInvoicesContainer = () => {
     setSendInvoiceModalOpen(false);
   };
 
-  const getFailedInvoicesData = (data: RECURRING_INVOICES | undefined) => {
-    if (!data?.berthLeases) return [];
-    return data.berthLeases.edges.reduce<FailedInvoices[]>((acc, edge) => {
-      if (!edge?.node) return acc;
-
-      const failedInvoice = {
-        id: edge.node.id,
-        customerId: edge.node.customer.id,
-        customerName: `${edge.node.customer.firstName} ${edge.node.customer.lastName}`,
-        harbor: edge.node.berth.pier.properties?.harbor.properties?.name ?? '',
-        failureReason: edge.node.comment,
-      };
-
-      return [...acc, failedInvoice];
-    }, []);
-  };
-
-  const getSummaryData = (data: RECURRING_INVOICES | undefined): DataSummaryProps['labelValuePairs'] => {
-    return [
-      { label: t('recurringInvoices.summary.customers'), value: data?.profiles?.count },
-      {
-        label: t('recurringInvoices.summary.sentSuccessfully'),
-        value: sendInvoicesData?.sendExistingBerthInvoices?.result?.successfulOrders?.length,
-      },
-      {
-        label: t('recurringInvoices.summary.failedInvoicing'),
-        value: data?.berthLeases?.count,
-      },
-    ];
-  };
+  const summaryData = [
+    { label: t('recurringInvoices.summary.customers'), value: data?.profiles?.count },
+    {
+      label: t('recurringInvoices.summary.sentSuccessfully'),
+      value: sendInvoicesData?.sendExistingBerthInvoices?.result?.successfulOrders?.length,
+    },
+    {
+      label: t('recurringInvoices.summary.failedInvoicing'),
+      value: data?.berthLeases?.count,
+    },
+  ];
 
   return (
     <>
       <RecurringInvoices
         loading={loading}
-        dataSummary={getSummaryData(data)}
+        dataSummary={summaryData}
         failedInvoicesCount={data?.berthLeases?.count}
         failedInvoicesData={getFailedInvoicesData(data)}
         handleSend={() => setSendInvoiceModalOpen(true)}
